NeuroAnalysis - Neural Signal Analysis

API Reference

NeuroAnalysis.CondDCh — Constant

Default digital input channels

NeuroAnalysis.SecondPerUnit — Constant

Time Unit set to millisecond by default.

NeuroAnalysis.badchmaskim — Method

Logical mask for IMEC bad and reference channels in probe shape

NeuroAnalysis.chmaskim — Method

Logical mask for IMEC channels in probe shape

NeuroAnalysis.circaxial — Function

Transforms p-axial data to a common scale α: sample of angles in radians p: number of modes

NeuroAnalysis.circaxialmean — Function

Mean direction for circular data with axial

NeuroAnalysis.circmean — Function

Mean for circular data

NeuroAnalysis.circr — Function

Mean Resultant Vector Length for circular data α: sample of angles in radians w: number of incidences in case of binned angle data d: spacing of bin centers for binned data, if supplied correction factor is used to correct for bias

NeuroAnalysis.circtuningstats — Method

Properties of Circular Tuning: Prefered Direction/Orientation Selectivity Index version 1: (ResponsePrefered - ResponseOpposing)/ResponsePrefered version 2: (ResponsePrefered - ResponseOpposing)/(ResponsePrefered + ResponseOpposing) Full Width at Half Maximum

x: angles in radius y: responses od: opposing angle distance, π for DSI, 0.5π for OSI s: symbol name

NeuroAnalysis.circvar — Function

Circular Variance

NeuroAnalysis.condfactor — Method

Get factors of conditions

NeuroAnalysis.condin — Method

Find unique condition and indices, repetition for each

NeuroAnalysis.condresponse — Method

Group repeats of Conditions, get Mean and SEM of responses

rs: responses of each trial gi: trial indices of repeats for each condition

NeuroAnalysis.condstring — Function

Condition in String

NeuroAnalysis.condtest — Method

Get CondTest DataFrame

NeuroAnalysis.condtestcond — Method

Get CondTestCond DataFrame

NeuroAnalysis.correlogram — Method

Shift(shuffle) corrected, normalized(coincidence/spike), condition/trial-averaged Cross-Correlogram of binary spike trains. (Bair, W., Zohary, E., and Newsome, W.T. (2001). Correlated Firing in Macaque Visual Area MT: Time Scales and Relationship to Behavior. J. Neurosci. 21, 1676–1697.)

NeuroAnalysis.csd — Method

1D Current Source Density for a set of voltage traces from a linear array of electrodes.

CSD: finite difference method of second spatial derivatives of potentials, I = - ∇σ⋅∇ϕ, σ is conductivities, ϕ is potentials, on the continuous assumption when spacing is around 20-100 um. (Nicholson & Freeman, 1975, J Neurophysiol, 38(2): 356-68)

iCSD: solve I -> ϕ forward problem, then get the inverse transformation matrix. (Petterson et al., 2006, J Neurosci Methods, 154(1-2):116-33)

  iCSDdelta: δ method where infinitely thin current source discs of radius r.
  iCSDspline: cubic spline method where current source between electrode be modeled by smooth spline

kCSD: kernal CSD ()

data: ch x sample matrix(volts) h: channel spacing(micrometers) c: conductivity of the extracellular medium in siemans/meter r: radius(micrometers) of the model current source/sink in reverse method

return: CSD(amps/meters^3)

NeuroAnalysis.csd_cubicspline — Function

cubic spline inverse CSD

contact_positions: contact positions data: measured potentials Fcs: cubic spline transformation matrix corresponding to the given contact positions E0,E1,E2,E3: matrixes containing the "recursive rules"

NeuroAnalysis.ctctc — Method

Get CondTest and CondTestCond DataFrame

NeuroAnalysis.digitalbit — Method

Parse bit event time and value.

NeuroAnalysis.digitaldata — Method

Get digital channel time and value.

NeuroAnalysis.dirsigtest — Method

Direction tuning significance using dot product with empirical orientation preference This function calculates the probability that the "true" direction tuning vector of a neuron has non-zero length. It performs this by empirically determing the unit orientation vector, and then computing the dot product of the direction vector for each trial onto the overall orientation vector, and then looking to see if the average is non-zero.

Inputs: ANGLES is a vector of direction angles at which the response has been measured. RATES is the response of the neuron in to each angle; each row should contain responses from a different trial. Output: P the probability that the "true" direction tuning vector is non-zero.

See: Mazurek, Kagan, Van Hooser 2014; Frontiers in Neural Circuits

NeuroAnalysis.dogf — Method

Difference of Gaussians function

NeuroAnalysis.e_matrixes — Method

Computes the E0, E1, E2 and E3 matrixes used in the cubic spline iCSD method. These matrixes contains the recursive formulas for finding the F matrix

NeuroAnalysis.expericafileregex — Method

Regular Expression to match Experica data file name

NeuroAnalysis.f_cubicspline — Function

F and E matrix of the cubic spline inverse CSD

NeuroAnalysis.factorresponse — Method

Condition Response in Factor Space

NeuroAnalysis.factorresponsestats — Method

Tuning properties of factor response

fl: factor levels fr: factor responses

HueAngle, Orientation and Direction follow the same convention such that 0 is -/→, then increase counter-clock wise.
For cases where Orientation and Direction are interlocked(drifting grating):
    when Orientation is -(0), then Direction is ↑(90)
    when Direction is →(0), then Orientation is |(-90)

NeuroAnalysis.finalfactor — Method

Get Final factors of conditions

NeuroAnalysis.flin — Method

Find levels for each factor and indices, repetition for each level

NeuroAnalysis.gaborf — Method

Gabor function

NeuroAnalysis.gaincorrectim — Method

Convert raw imec data to gain-corrected voltages. The saved-channel id subset in data will be used to get corresponding gain.

The full conversion with gain is: dataVolts = dataInt * fI2V / gain. Note that each channel may have its own gain.

y: the saved epoch of all imec channels, excluding the last Sync channel.

NeuroAnalysis.gaussianf — Method

Gaussian function

NeuroAnalysis.getdatafile — Method

Get matched file names in path

NeuroAnalysis.getexpericafile — Method

Get matched Experica files in directory

NeuroAnalysis.getoifile — Method

Get Optical Imaging VDAQ matched files in directory

NeuroAnalysis.grating — Method

Generate Grating Image.

θ: Orientation (radius)

sf: SpatialFreq (cycle/deg)

phase: Sin wave phase in [0,1]

stisize: Tuple of image size in visual degree

ppd: Pixel per degree

return image in [0,1]

NeuroAnalysis.gratingf — Method

Plane sin wave funtion

NeuroAnalysis.gvmf — Function

Generalized von Mises function [1]

\[f(α) = βe^{κ₁(cos(α - μ₁) - 1) + κ₂(cos2(α - μ₂) - 1)}\]

[1]

Gatto, R., and Jammalamadaka, S.R. (2007). The generalized von Mises distribution. Statistical Methodology 4, 341–353.

NeuroAnalysis.hlpass — Method

High pass and low pass filters

NeuroAnalysis.hotellingt2test — Function

Hotelling's T-Squared test for one multivariate sample.

Performs Hotelling's T^2 test on multivariate samples X to determine if the data have mean MU. X should be a NxP matrix with N observations of P-dimensional data, and the mean MU to be tested should be 1xP. the significance level α is 0.05 by default.

H is 1 if the null hypothesis (that the mean of X is equal to MU) can be rejected at significance level α. P is the actual P value.

The code is based on HotellingT2.m by A. Trujillo-Ortiz and R. Hernandez-Walls.

NeuroAnalysis.ismodulative — Method

Check if any factors and their interactions significently modulate response by ANOVA

NeuroAnalysis.isresponsive — Method

Check if any sub group of response is significently different from baseline by Wilcoxon Signed Rank Test

NeuroAnalysis.isresponsive — Method

Check if response is significently different from baseline by Wilcoxon Signed Rank Test

NeuroAnalysis.loadimageset — Method

Load images in directory

NeuroAnalysis.maptodatatime — Method

Map time to data timeline, optionally add latency.

NeuroAnalysis.mat2julia! — Method

Convert MATLAB type to Julia type with proper dimention

NeuroAnalysis.matchfile — Method

Get matched files in directory, optionally add directory to get file path

NeuroAnalysis.matdictarray2df — Method

convert Matlab struct of array to DataFrame

NeuroAnalysis.oifileregex — Method

Regular Expression to match Optical Imaging VDAQ block file name

NeuroAnalysis.parsedigitalinanalog — Function

Get digital rising and falling edges in a analog stream return di: edges index dv: edges active/inactive state

NeuroAnalysis.poissonspike — Method

Generate Poisson Spike Train

NeuroAnalysis.prepare — Method

Prepare Dataset in MATLAB MAT File

NeuroAnalysis.prepare_experica! — Method

Prepare Experica Experiment Data

NeuroAnalysis.prepare_oi! — Method

Prepare Optical Imaging Block Data

NeuroAnalysis.prepare_ripple! — Method

Prepare Ripple Data

NeuroAnalysis.readmat — Method

Read variables in MATLAB MAT file

NeuroAnalysis.readmeta — Method

Read DataExport Metadata MAT File

NeuroAnalysis.refchmaskim — Method

Logical mask for IMEC reference channels in probe shape

NeuroAnalysis.rmline! — Method

Remove line noise and its harmonics by notch filter

NeuroAnalysis.roccurve — Method

This function calculates the ROC curve, which represents the 1-specificity and sensitivity of two classes of data, (i.e., class1 and class2).

The function also returns all the needed quantitative parameters: threshold position, distance to the optimum point, sensitivity, specificity, accuracy, area under curve (AROC), positive and negative predicted values (PPV, NPV), false negative and positive rates (FNR, FPR), false discovery rate (FDR), false omission rate (FOR), F1 score, Matthews correlation coefficient (MCC), Informedness (BM) and Markedness; as well as the number of true positives (TP), true negatives (TN), false positives (FP), and false negatives (FN).

Rewrite from Víctor Martínez-Cagigal (2020). ROC Curve, MATLAB Central File Exchange. Retrieved February 11, 2020.

NeuroAnalysis.sta — Method

Spike Triggered Average of Images

x: Matrix where each row is one image

y: Vector of image response

NeuroAnalysis.statetime — Method

Extract state time of condtest

NeuroAnalysis.subrv — Method

Sub sets of RealVector in between binedges

NeuroAnalysis.subrv — Method

Sub set of a RealVector rv, where min <= rv[i] < max, kwargs isminzero, ismaxzero and shift set subrv zero value to min+shift or max+shift.

return:

y: subset of RealVector

n: number of elements in the subset, or divided by duration(max-min) of the subset, based on kwargs israte

w: window of the subset (min,max)

i: indices of subset elements in original RealVector, such that y = rv[i]

See also: subrvr, histrv

NeuroAnalysis.subrvr — Method

Response of each sub set of RealVector, could be mean firing rate or number of spikes based on kwarg israte.

See also: subrvr_ono

NeuroAnalysis.subrvr_ono — Method

Response of each sub set of RealVector, could be mean firing rate or number of spikes based on kwarg israte.

Note

This is a faster(~900x) version compared to subrvr, but only works when rv, mins and maxs are ascendingly ordered, and each maxs-mins range are none-overlapping.

NeuroAnalysis.unitfyspike_kilosort — Method

Organize each spiking unit info from Kilosort result.

templateradius: radius(μm) within which the templates height are used to estimate unit position.

NeuroAnalysis.vlabregex — Method

Regular expression to match VLab data file names

NeuroAnalysis.vmf — Function

von Mises function [1]

\[f(α) = βe^{κ(cos(n(α - μ)) - 1)}\]

[1]

Swindale, N.V. (1998). Orientation tuning curves: empirical description and estimation of parameters. Biol Cybern 78, 45–56.

β: amplitude at μ

μ: angle of peak

κ: width parameter

n: frequency parameter

Index

NeuroAnalysis.CondDCh

NeuroAnalysis.SecondPerUnit

NeuroAnalysis.badchmaskim

NeuroAnalysis.chmaskim

NeuroAnalysis.circaxial

NeuroAnalysis.circaxialmean

NeuroAnalysis.circmean

NeuroAnalysis.circr

NeuroAnalysis.circtuningstats

NeuroAnalysis.circvar

NeuroAnalysis.condfactor

NeuroAnalysis.condin

NeuroAnalysis.condresponse

NeuroAnalysis.condstring

NeuroAnalysis.condtest

NeuroAnalysis.condtestcond

NeuroAnalysis.correlogram

NeuroAnalysis.csd

NeuroAnalysis.csd_cubicspline

NeuroAnalysis.ctctc

NeuroAnalysis.digitalbit

NeuroAnalysis.digitaldata

NeuroAnalysis.dirsigtest

NeuroAnalysis.dogf

NeuroAnalysis.e_matrixes

NeuroAnalysis.expericafileregex

NeuroAnalysis.f_cubicspline

NeuroAnalysis.factorresponse

NeuroAnalysis.factorresponsestats

NeuroAnalysis.finalfactor

NeuroAnalysis.flin

NeuroAnalysis.gaborf

NeuroAnalysis.gaincorrectim

NeuroAnalysis.gaussianf

NeuroAnalysis.getdatafile

NeuroAnalysis.getexpericafile

NeuroAnalysis.getoifile

NeuroAnalysis.grating

NeuroAnalysis.gratingf

NeuroAnalysis.gvmf

NeuroAnalysis.hlpass

NeuroAnalysis.hotellingt2test

NeuroAnalysis.ismodulative

NeuroAnalysis.isresponsive

NeuroAnalysis.isresponsive

NeuroAnalysis.loadimageset

NeuroAnalysis.maptodatatime

NeuroAnalysis.mat2julia!

NeuroAnalysis.matchfile

NeuroAnalysis.matdictarray2df

NeuroAnalysis.oifileregex

NeuroAnalysis.parsedigitalinanalog

NeuroAnalysis.poissonspike

NeuroAnalysis.prepare

NeuroAnalysis.prepare_experica!

NeuroAnalysis.prepare_oi!

NeuroAnalysis.prepare_ripple!

NeuroAnalysis.readmat

NeuroAnalysis.readmeta

NeuroAnalysis.refchmaskim

NeuroAnalysis.rmline!

NeuroAnalysis.roccurve

NeuroAnalysis.sta

NeuroAnalysis.statetime

NeuroAnalysis.subrv

NeuroAnalysis.subrv

NeuroAnalysis.subrvr

NeuroAnalysis.subrvr_ono

NeuroAnalysis.unitfyspike_kilosort

NeuroAnalysis.vlabregex

NeuroAnalysis.vmf